+++ test/unit/slurm/test_slurm_config_generator.py
@@ -11,16 +11,18 @@
 
 import inspect
 import os
-import pytest
-import slurm
-from assertpy import assert_that
 from pathlib import Path
+
+import pytest
+from assertpy import assert_that
 from pcluster_slurm_config_generator import generate_slurm_config_files
+
 
 def _get_template_folder() -> str:
     module_file = inspect.getsourcefile(generate_slurm_config_files)
     template_path = Path(module_file).parent / "templates"
     return str(template_path)
+
 
 @pytest.mark.parametrize(
     "no_gpu",
@@ -31,9 +33,7 @@ def test_generate_slurm_config_files_nogpu(mocker, test_datadir, tmpdir, no_gpu)
     instance_types_data = str(test_datadir / "sample_instance_types_data.json")
 
     mocker.patch("pcluster_slurm_config_generator.gethostname", return_value="ip-1-0-0-0", autospec=True)
-    mocker.patch(
-        "pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True
-    )
+    mocker.patch("pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True)
     template_directory = _get_template_folder()
     generate_slurm_config_files(
         tmpdir,
@@ -74,9 +74,7 @@ def test_generate_slurm_config_files_memory_scheduling(
     instance_types_data = str(test_datadir / "sample_instance_types_data.json")
 
     mocker.patch("pcluster_slurm_config_generator.gethostname", return_value="ip-1-0-0-0", autospec=True)
-    mocker.patch(
-        "pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True
-    )
+    mocker.patch("pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True)
     template_directory = _get_template_folder()
     generate_slurm_config_files(
         tmpdir,
@@ -118,9 +116,7 @@ def test_generate_slurm_config_files_slurm_accounting(mocker, test_datadir, tmpd
     instance_types_data = str(test_datadir / "sample_instance_types_data.json")
 
     mocker.patch("pcluster_slurm_config_generator.gethostname", return_value="ip-1-0-0-0", autospec=True)
-    mocker.patch(
-        "pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True
-    )
+    mocker.patch("pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True)
     template_directory = _get_template_folder()
     generate_slurm_config_files(
         tmpdir,
@@ -144,9 +140,7 @@ def test_generate_slurm_config_files_slurm_accounting(mocker, test_datadir, tmpd
 
 def test_generating_slurm_config_flexible_instance_types(mocker, test_datadir, tmpdir):
     mocker.patch("pcluster_slurm_config_generator.gethostname", return_value="ip-1-0-0-0", autospec=True)
-    mocker.patch(
-        "pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True
-    )
+    mocker.patch("pcluster_slurm_config_generator._get_head_node_private_ip", return_value="ip.1.0.0.0", autospec=True)
 
     input_file = os.path.join(test_datadir, "sample_input.yaml")
     instance_types_data = os.path.join(test_datadir, "sample_instance_types_data.json")
